Add some related items
I've been wanting to be able to link things together from the very
beginning. It was going to be one of the main advantages of having a
system like this. You can always link things together to create a whole
network of information. That idea is one of the reasons I started.